xueqiu/requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
astroid==2.7.3
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
ddt==1.4.2
execnet==1.9.0
idna==3.2
importlib-metadata==4.8.1
iniconfig==1.1.1
isort==5.9.3
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
packaging==21.0
platformdirs==2.3.0
pluggy==1.0.0
py==1.10.0
pylint==2.10.2
pyparsing==2.4.7
pytest==6.2.5
pytest-assume==2.4.3
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-rerunfailures==10.2
pytest-xdist==2.4.0
PyYAML==5.4.1
requests==2.26.0
selenium==3.141.0
six==1.16.0
toml==0.10.2
tqdm==4.62.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
wrapt==1.12.1
zipp==3.6.0
